+++ src/Constants/CaseTypes.php
@@ -2,19 +2,29 @@
 
 declare(strict_types=1);
 
-namespace SebastiaanLuca\AutoMorphMap;
+namespace SebastiaanLuca\AutoMorphMap\Constants;
 
 class CaseTypes
 {
     /**
      * @var string
      */
-    public const CAMEL_CASE = 'camelCase';
+    public const NONE = 'none';
 
     /**
      * @var string
      */
     public const SNAKE_CASE = 'snake_case';
+
+    /**
+     * @var string
+     */
+    public const SLUG_CASE = 'slug-case';
+
+    /**
+     * @var string
+     */
+    public const CAMEL_CASE = 'camelCase';
 
     /**
      * @var string