+++ tests/Feature/AutobinderCacheTest.php
@@ -25,11 +25,11 @@ class AutobinderCacheTest extends TestCase
             'user' => 'App\\User',
             'something_inherited' => 'App\\Models\\SomethingInherited',
             'address' => 'MyModule\\Models\\Address',
-            'thing' => 'MyPackage\\Models\\Thing',
-            'package' => 'MyPackage\\Models\\Sub\\Package',
+            'SomeThing' => 'MyPackage\\Models\\Thing',
+            'different_package' => 'MyPackage\\Models\\Sub\\Package',
         ];
 
-        $relation->shouldReceive('morphMap')->once();
+        $relation->shouldReceive('morphMap')->once()->withNoArgs();
         $relation->shouldReceive('morphMap')->once()->with($expected);
 
         app(Mapper::class)->map();
@@ -43,7 +43,7 @@ class AutobinderCacheTest extends TestCase
         $cache = base_path('bootstrap/cache/morphmap.php');
 
         $copy = copy(
-            base_path('cache.php'),
+            __DIR__ . '/../resources/cache.php',
             $cache
         );
 
@@ -52,11 +52,11 @@ class AutobinderCacheTest extends TestCase
         $relation = $this->getMockedRelation();
 
         $expected = [
-            'something_cached' => 'App\\SomethingCached',
-            'cached_user' => 'App\\Models\\CachedUser',
+            'something_inherited' => 'App\\Models\\SomethingInherited',
+            'different_package' => 'MyPackage\\Models\\Sub\\Package',
         ];
 
-        $relation->shouldReceive('morphMap')->once();
+        $relation->shouldReceive('morphMap')->once()->withNoArgs();
         $relation->shouldReceive('morphMap')->once()->with($expected);
 
         app(Mapper::class)->map();