+++ tests/Feature/AutobinderNamingTest.php
@@ -6,23 +6,23 @@ namespace SebastiaanLuca\AutoMorphMap\Tests\Feature;
 
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Mockery\MockInterface;
-use SebastiaanLuca\AutoMorphMap\Constants\CaseTypes;
+use SebastiaanLuca\AutoMorphMap\Constants\NamingSchemes;
 use SebastiaanLuca\AutoMorphMap\Mapper;
 use SebastiaanLuca\AutoMorphMap\Tests\MocksInstances;
 use SebastiaanLuca\AutoMorphMap\Tests\TestCase;
 
-class AutobinderCaseTest extends TestCase
+class AutobinderNamingTest extends TestCase
 {
     use MocksInstances;
 
     /**
      * @test
      */
-    public function it maps all models using the default case() : void
+    public function it maps all models using the default naming scheme() : void
     {
         $relation = $this->getMockedRelation();
 
-        config()->set('auto-morph-map.case', null);
+        config()->set('auto-morph-map.naming', null);
 
         $expected = [
             'user' => 'App\\User',
@@ -41,17 +41,17 @@ class AutobinderCaseTest extends TestCase
     /**
      * @test
      */
-    public function it maps all models using snake case() : void
+    public function it maps all models using the singular table name naming scheme() : void
     {
         $relation = $this->getMockedRelation();
 
-        config()->set('auto-morph-map.case', CaseTypes::SNAKE_CASE);
+        config()->set('auto-morph-map.naming', NamingSchemes::SINGULAR_TABLE_NAME);
 
         $expected = [
             'user' => 'App\\User',
-            'something_inherited' => 'App\\Models\\SomethingInherited',
+            'something_inherited' => 'App\Models\\SomethingInherited',
             'address' => 'MyModule\\Models\\Address',
-            'some_thing' => 'MyPackage\\Models\\Thing',
+            'SomeThing' => 'MyPackage\\Models\\Thing',
             'different_package' => 'MyPackage\\Models\\Sub\\Package',
         ];
 
@@ -64,18 +64,18 @@ class AutobinderCaseTest extends TestCase
     /**
      * @test
      */
-    public function it maps all models using slug case() : void
+    public function it maps all models using the table name naming scheme() : void
     {
         $relation = $this->getMockedRelation();
 
-        config()->set('auto-morph-map.case', CaseTypes::SLUG_CASE);
+        config()->set('auto-morph-map.naming', NamingSchemes::TABLE_NAME);
 
         $expected = [
-            'user' => 'App\\User',
-            'something-inherited' => 'App\\Models\\SomethingInherited',
-            'address' => 'MyModule\\Models\\Address',
-            'something' => 'MyPackage\\Models\\Thing',
-            'different-package' => 'MyPackage\\Models\\Sub\\Package',
+            'users' => 'App\\User',
+            'something_inheriteds' => 'App\Models\\SomethingInherited',
+            'addresses' => 'MyModule\\Models\\Address',
+            'SomeThings' => 'MyPackage\\Models\\Thing',
+            'different_packages' => 'MyPackage\\Models\\Sub\\Package',
         ];
 
         $relation->shouldReceive('morphMap')->once()->withNoArgs();
@@ -87,41 +87,18 @@ class AutobinderCaseTest extends TestCase
     /**
      * @test
      */
-    public function it maps all models using camel case() : void
+    public function it maps all models using the class basename naming scheme() : void
     {
         $relation = $this->getMockedRelation();
 
-        config()->set('auto-morph-map.case', CaseTypes::CAMEL_CASE);
-
-        $expected = [
-            'user' => 'App\\User',
-            'somethingInherited' => 'App\\Models\\SomethingInherited',
-            'address' => 'MyModule\\Models\\Address',
-            'someThing' => 'MyPackage\\Models\\Thing',
-            'differentPackage' => 'MyPackage\\Models\\Sub\\Package',
-        ];
-
-        $relation->shouldReceive('morphMap')->once()->withNoArgs();
-        $relation->shouldReceive('morphMap')->once()->with($expected);
-
-        app(Mapper::class)->map();
-    }
-
-    /**
-     * @test
-     */
-    public function it maps all models using studly case() : void
-    {
-        $relation = $this->getMockedRelation();
-
-        config()->set('auto-morph-map.case', CaseTypes::STUDLY_CASE);
+        config()->set('auto-morph-map.naming', NamingSchemes::CLASS_BASENAME);
 
         $expected = [
             'User' => 'App\\User',
-            'SomethingInherited' => 'App\\Models\\SomethingInherited',
+            'SomethingInherited' => 'App\Models\\SomethingInherited',
             'Address' => 'MyModule\\Models\\Address',
-            'SomeThing' => 'MyPackage\\Models\\Thing',
-            'DifferentPackage' => 'MyPackage\\Models\\Sub\\Package',
+            'Thing' => 'MyPackage\\Models\\Thing',
+            'Package' => 'MyPackage\\Models\\Sub\\Package',
         ];
 
         $relation->shouldReceive('morphMap')->once()->withNoArgs();