+++ tests/Feature/MapperTest.php
@@ -25,11 +25,11 @@ class MapperTest extends TestCase
             'user' => 'App\\User',
             'something_inherited' => 'App\\Models\\SomethingInherited',
             'address' => 'MyModule\\Models\\Address',
-            'thing' => 'MyPackage\\Models\\Thing',
-            'package' => 'MyPackage\\Models\\Sub\\Package',
+            'SomeThing' => 'MyPackage\\Models\\Thing',
+            'different_package' => 'MyPackage\\Models\\Sub\\Package',
         ];
 
-        $relation->shouldReceive('morphMap')->once();
+        $relation->shouldReceive('morphMap')->once()->withNoArgs();
         $relation->shouldReceive('morphMap')->once()->with($expected);
 
         app(Mapper::class)->map();
@@ -50,10 +50,10 @@ class MapperTest extends TestCase
         $expected = [
             'user' => 'App\\User',
             'address' => 'MyModule\\Models\\Address',
-            'package' => 'MyPackage\\Models\\Sub\\Package',
+            'different_package' => 'MyPackage\\Models\\Sub\\Package',
         ];
 
-        $relation->shouldReceive('morphMap')->once()->andReturn($existing);
+        $relation->shouldReceive('morphMap')->once()->withNoArgs()->andReturn($existing);
         $relation->shouldReceive('morphMap')->once()->with($expected);
 
         app(Mapper::class)->map();